code cleaning

+++ xfl/xflwidgets/customdlg/doublevaluedlg.h
@@ -21,7 +21,7 @@ class DoubleValueDlg : public XflDialog
     public:
         DoubleValueDlg(QWidget *pParent, QVector<double> values, QStringList const &leftlabels, QStringList const &rightlabels);
 
-        void setValue(int iVal, int value) {if(iVal>=0 && iVal<m_pDoubleEdit.size()) m_pDoubleEdit[iVal]->setValue(value);}
+        void setValue(int iVal, double value) {if(iVal>=0 && iVal<m_pDoubleEdit.size()) m_pDoubleEdit[iVal]->setValue(value);}
         double value(int iVal) const{if(iVal>=0 && iVal<m_pDoubleEdit.size()) return m_pDoubleEdit[iVal]->value(); else return 0;}
 
     private:
